Form Validation added

+++ lib/auth/login.dart
@@ -14,10 +14,10 @@ class Login extends StatefulWidget {
 
 class _LoginState extends State<Login> {
   final _formKey = GlobalKey<FormState>();
-  String _message = ' ';
   Agency _user = Agency();
   bool _invisiblePwd = true;
   bool _isLoading = false;
+  bool _autoValidate = false;
 
   @override
   Widget build(BuildContext context) {
@@ -50,16 +50,9 @@ class _LoginState extends State<Login> {
               Padding(
                 padding: EdgeInsets.all(32.0),
                 child: Center(
-                  child: _isLoading
-                      ? LinearProgressIndicator()
-                      : Text(
-                          this._message,
-                          style: TextStyle(
-                            color: Colors.red[400],
-                            fontSize: 16.0,
-                          ),
-                        ),
-                ),
+                    child: _isLoading
+                        ? LinearProgressIndicator()
+                        : SizedBox(height: 6.0)),
               ),
               Builder(
                 builder: (context) => Form(
@@ -71,20 +64,12 @@ class _LoginState extends State<Login> {
                       children: <Widget>[
                         TextFormField(
                           enabled: !this._isLoading,
-                          decoration: InputDecoration(
-                            enabledBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: Colors.teal[100],
-                              ),
-                            ),
-                            focusedBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: Colors.teal,
-                              ),
-                            ),
-                            labelText: 'Licence',
-                            prefixIcon: Icon(Icons.credit_card),
-                          ),
+                          decoration:
+                              decorate("License", Icons.credit_card, false),
+                          validator: (value) {
+                            return validate(value);
+                          },
+                          autovalidate: _autoValidate,
                           onSaved: (val) => _user.licence = val,
                         ),
                         SizedBox(
@@ -92,34 +77,11 @@ class _LoginState extends State<Login> {
                         ),
                         TextFormField(
                           enabled: !this._isLoading,
-                          decoration: InputDecoration(
-                            labelText: 'Password',
-                            prefixIcon: Icon(
-                              Icons.vpn_key,
-                            ),
-                            suffixIcon: IconButton(
-                                color: Colors.black54,
-                                padding: EdgeInsets.all(1.0),
-                                icon: this._invisiblePwd
-                                    ? Icon(Icons.visibility)
-                                    : Icon(Icons.visibility_off),
-                                onPressed: () {
-                                  setState(() {
-                                    this._invisiblePwd = !this._invisiblePwd;
-                                  });
-                                }),
-                            enabledBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: Colors.teal[100],
-                              ),
-                            ),
-                            focusColor: Colors.blue,
-                            focusedBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: Colors.blue,
-                              ),
-                            ),
-                          ),
+                          decoration: decorate("Password", Icons.vpn_key, true),
+                          validator: (value) {
+                            return validate(value);
+                          },
+                          autovalidate: _autoValidate,
                           onSaved: (val) => setState(() => _user.psswd = val),
                           obscureText: this._invisiblePwd,
                         ),
@@ -136,8 +98,6 @@ class _LoginState extends State<Login> {
                       onPressed: this._isLoading
                           ? null
                           : () {
-                              final form = _formKey.currentState;
-                              form.save();
                               handleLogin(context);
                             },
                     ),
@@ -154,27 +114,34 @@ class _LoginState extends State<Login> {
   handleLogin(cntxt) async {
     String mssg;
     Color color;
-    setState(() {
-      this._isLoading = true;
-    });
-    Map<String, dynamic> result = await userSignin(_user);
-    if (result['auth_token'] != null) {
-      mssg = 'Welcome Back';
-      color = Colors.green[400];
+
+    if (_formKey.currentState.validate()) {
+      _formKey.currentState.save();
       setState(() {
-        this._isLoading = false;
+        this._isLoading = true;
       });
-      await storage.write(key: 'auth_token', value: result['auth_token']);
-      await storage.write(key: 'prompted', value: 'true');
-      Navigator.pop(context);
-      Navigator.pushReplacementNamed(context, '/home');
+      Map<String, dynamic> result = await userSignin(_user);
+      if (result['auth_token'] != null) {
+        mssg = 'Welcome Back';
+        color = Colors.green[400];
+        setState(() {
+          this._isLoading = false;
+        });
+        await storage.write(key: 'auth_token', value: result['auth_token']);
+        await storage.write(key: 'prompted', value: 'true');
+        Navigator.pop(context);
+        Navigator.pushReplacementNamed(context, '/home');
+      } else {
+        mssg = result['err'];
+        color = Colors.red[400];
+        setState(() {
+          this._isLoading = false;
+        });
+        storage.deleteAll();
+      }
     } else {
-      mssg = result['err'];
+      mssg = "Please fill up all the details";
       color = Colors.red[400];
-      setState(() {
-        this._isLoading = false;
-      });
-      storage.deleteAll();
     }
 
     final SnackBar snackBar = msgBar(
@@ -182,5 +149,58 @@ class _LoginState extends State<Login> {
       color,
     );
     Scaffold.of(cntxt).showSnackBar(snackBar);
+  }
+
+  String validate(value) {
+    this._autoValidate = true;
+    if (value.isEmpty) {
+      return 'Please fill this field';
+    }
+    return null;
+  }
+
+  InputDecoration decorate(String text, IconData icon, bool password) {
+    return InputDecoration(
+      enabledBorder: OutlineInputBorder(
+        borderSide: BorderSide(
+          color: Colors.teal[100],
+        ),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderSide: BorderSide(
+          color: Colors.green,
+        ),
+      ),
+      disabledBorder: OutlineInputBorder(
+        borderSide: BorderSide(
+          color: Colors.grey,
+        ),
+      ),
+      errorBorder: OutlineInputBorder(
+        borderSide: BorderSide(
+          color: Colors.red,
+        ),
+      ),
+      focusedErrorBorder: OutlineInputBorder(
+        borderSide: BorderSide(
+          color: Colors.deepOrange[200],
+        ),
+      ),
+      labelText: text,
+      prefixIcon: Icon(icon),
+      suffixIcon: password
+          ? IconButton(
+              color: Colors.black54,
+              padding: EdgeInsets.all(1.0),
+              icon: this._invisiblePwd
+                  ? Icon(Icons.visibility)
+                  : Icon(Icons.visibility_off),
+              onPressed: () {
+                setState(() {
+                  this._invisiblePwd = !this._invisiblePwd;
+                });
+              })
+          : null,
+    );
   }
 }
